Split search into multiple files.
