+++ app/search/index.py
@@ -1,76 +1,19 @@
 
 """
-Build a search index from the database.
+Build a search index of the database.
 """
 
-# Possible improvements
-#   > Turn words into their lexemes to handle plurality and tenses of words.
-#   > Promote search results where search terms show up together in the same 
-#     order.
-#   > Weight searches based on word hit counts and other stats.
-#   > Throw away stop words (and, or, with, of)
-#   > Handle unicode characters.
-
 import os
-import sys
-import time
 import re
 import pickle
 from pathlib import Path
-from app.api.database_connector import database_connect
+import app.search.descriptions as desc
 
-# TODO: This does not belong here.
-def format_minutes(ready_time):
+def download_descriptions(db):
     """
-    Take an integral amount of minutes and return a human-readable text version.
+    Download database data, generate descriptions, and save the descriptions
+    locally.
     """
-
-    minutes = ready_time % 60
-    hours = (ready_time // 60) % 24
-    days = ready_time // (60 * 24)
-
-    unit_strs = []
-
-    if days == 1:
-        unit_strs.append("{} day".format(days))
-    elif days > 1:
-        unit_strs.append("{} days".format(days))
-
-    if hours == 1:
-        unit_strs.append("{} hour".format(hours))
-    elif hours > 1:
-        unit_strs.append("{} hours".format(hours))
-
-    if minutes == 1:
-        unit_strs.append("{} minute".format(minutes))
-    elif minutes > 1:
-        unit_strs.append("{} minutes".format(minutes))
-
-    return ", ".join(unit_strs)
-
-# TODO: This does not belong here.
-def describe_recipe(recipe):
-    """
-    Generate a text description of a recipe's attributes.
-    """
-
-    fmt = ("{name}\n"
-           "Recipe id: {recipe_id}\n"
-           "Servings: {servings}\n"
-           "Ready in: {readyInMinutes}\n"
-           "Decription: {description}\n"
-           "Instructions: {instructions}")
-
-    return fmt.format(name=recipe.name,
-                      recipe_id=recipe.recipe_id,
-                      servings=recipe.servings,
-                      readyInMinutes=format_minutes(recipe.ready_time),
-                      description=recipe.description,
-                      instructions=recipe.instructions)
-
-
-
-def cmd_text_db(db):
 
     res = db.engine.execute("SELECT count(recipe_id) as cnt FROM recipe;")
     recipe_count = res.fetchone().cnt
@@ -85,7 +28,7 @@ def cmd_text_db(db):
                                 "OFFSET {index} LIMIT 1;".format(index=index))
 
         recipe = res.fetchone()
-        text = describe_recipe(recipe)
+        text = desc.describe_recipe(recipe)
 
         path = "data/recipes/{}.txt".format(recipe.recipe_id)
         os.makedirs(os.path.dirname(path), exist_ok=True)
@@ -95,10 +38,7 @@ def cmd_text_db(db):
         if index % status_freq == 0:
             print("Progress: {:.1f}%".format((index / recipe_count) * 100))
 
-def cmd_text(args):
-    database_connect(cmd_text_db)
-
-def cmd_build(args):
+def build_index():
     index = dict()
 
     # TODO: Move constant to an appropriate location.
@@ -137,200 +77,4 @@ def cmd_build(args):
     # pillars.
     pickle.dump(index, open("index.p", "wb"))
 
-class SearchResult:
-
-    def __init__(self, pillar, item_id, description, terms):
-        self.pillar = pillar
-        self.item_id = item_id
-        self.description = description
-        self.terms = terms
-
-def split_query(query):
-    return list(re.compile("([^\s]+)").findall(query))
-
-def contextualize(description, query):
-    args = split_query(query)
-
-    # TODO: Search contextualization.
-
-    return [description]
-
-def search(query, page_number, page_size):
-    """
-    Performs a search on all models and their attributes. Returns a list of
-    SearchResult objects.
-    """
-
-    args = split_query(query)
-
-    # TODO: Don't load pickle file, use database instead.
-    try:
-        start = time.perf_counter()
-        # Loading the index takes a few seconds usually.
-        index = pickle.load(open("index.p", "rb"))
-
-        timediff = time.perf_counter() - start
-        print("Index loaded in {:.4f} seconds.".format(timediff))
-    except FileNotFoundError as error:
-        print("Index file index.p not found. You need to build the index"
-              " first.\n"
-              "\tpython index.py text\n"
-              "\tpython index.py build\n")
-        return
-
-    start_time = time.perf_counter()
-
-    # Build a dictionary mapping recipes to a set of terms they contain.
-    recipe_termset = dict()
-    for term in args:
-        if term not in index:
-            continue
-
-        for recipe_id in index[term]:
-            if recipe_id not in recipe_termset:
-                recipe_termset[recipe_id] = set([term])
-            else:
-                recipe_termset[recipe_id].add(term)
-
-    # No results found, exit early.
-    if not recipe_termset:
-        timediff = time.perf_counter() - start
-        print("No results found. Took {:.6f} seconds".format(timediff))
-
-    # Flip recipe_termset to get a termset -> recipes dictionary.
-    termset_recipes = dict()
-    for recipe_id in recipe_termset:
-        terms = tuple(recipe_termset[recipe_id])
-        if terms not in termset_recipes:
-            termset_recipes[terms] = set([recipe_id])
-        else:
-            termset_recipes[terms].add(recipe_id)
-
-    # Display and count the results.
-
-    start = page_number * page_size
-    end = start + page_size
-    set_n = 0
-    total_index = 0
-    search_results = []
-    sorted_keys = sorted(termset_recipes.keys(),
-                         key=lambda tup: len(tup),
-                         reverse=True)
-
-    # Find the first result set for the page.
-    while set_n < len(termset_recipes):
-        termset_results = termset_recipes[sorted_keys[set_n]]
-        if total_index + len(termset_results) > start:
-            break
-        total_index += len(termset_results)
-        set_n += 1
-
-    result_index = start - total_index
-    while (len(search_results) < end - start
-           and set_n < len(termset_recipes)):
-        termset_results = list(termset_recipes[sorted_keys[set_n]])
-        while (len(search_results) < end - start
-               and result_index < len(termset_results)):
-            search_results.append(termset_results[result_index])
-            result_index += 1
-        result_index = 0
-        set_n += 1
-
-    return search_results
-
-def cmd_fullsearch(args):
-
-    assert(len(args) >= 3)
-
-    start = int(args[0])
-    end = int(args[1])
-
-    print(search(" ".join(args[2:]), start, end))
-
-def cmd_search(args):
-
-    if len(args) == 0:
-        print("No search terms given.")
-        return
-
-    index = None
-    try:
-        start = time.perf_counter()
-        # Loading the index takes a few seconds usually.
-        index = pickle.load(open("index.p", "rb"))
-
-        timediff = time.perf_counter() - start
-        print("Index loaded in {:.4f} seconds.".format(timediff))
-    except FileNotFoundError as error:
-        print("Index file index.p not found. You need to build the index"
-              " first.\n"
-              "\tpython index.py text\n"
-              "\tpython index.py build\n")
-        return
-
-    start_time = time.perf_counter()
-
-    # Build a dictionary mapping recipes to a set of terms they contain.
-    recipe_termset = dict()
-    for term in args:
-        if term not in index:
-            continue
-
-        for recipe_id in index[term]:
-            if recipe_id not in recipe_termset:
-                recipe_termset[recipe_id] = set([term])
-            else:
-                recipe_termset[recipe_id].add(term)
-
-    # No results found, exit early.
-    if not recipe_termset:
-        timediff = time.perf_counter() - start
-        print("No results found. Took {:.6f} seconds".format(timediff))
-
-    # Flip recipe_termset to get a termset -> recipes dictionary.
-    termset_recipes = dict()
-    for recipe_id in recipe_termset:
-        terms = tuple(recipe_termset[recipe_id])
-        if terms not in termset_recipes:
-            termset_recipes[terms] = set([recipe_id])
-        else:
-            termset_recipes[terms].add(recipe_id)
-
-    # Display and count the results.
-    results_count = 0
-    for termset in sorted(termset_recipes.keys(),
-                              key=lambda tup: len(tup),
-                              reverse=True):
-        print(termset)
-        print(termset_recipes[termset])
-        print("\n")
-        results_count += len(termset_recipes[termset])
-
-    timediff = time.perf_counter() - start
-
-    print("{num_results} results found in {seconds:.6f} seconds.\n"
-          .format(num_results=results_count,
-                  seconds=timediff))
-
-def main():
-
-    commands = {"text": cmd_text,
-                "build": cmd_build,
-                "search": cmd_search,
-                "fullsearch": cmd_fullsearch}
-
-    if len(sys.argv) <= 1 or sys.argv[1] not in commands:
-        print("Possible commands are:\n"
-              "python index.py text  - generate text versions of database"
-              " items and save them locally.\n"
-              "python index.py build - builds the search index cache."
-              "python index.py search TERMS - conduct a search using a the"
-              " index cache.\n"
-              "python index.py search PAGE PAGESIZE TERMS - conduct a search"
-              " with pagination and contextualization.")
-    else:
-        commands[sys.argv[1]](sys.argv[2:])
-
-if __name__ == "__main__":
-    main()
